requirements.txt
scikit-learn
pytorch
torchvision
tensorboardX
